refactor: split tsc functionality into separate files
